pyproject.toml
[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra --strict-markers"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[project]
name = "rylan-unifi-case-study"
version = "0.1.0"
description = "UniFi + AI helpdesk case study infrastructure-as-code"
requires-python = ">=3.11"
dependencies = [
	"requests>=2.32.0,<3",
	"PyYAML>=6.0.0,<7",
	"pydantic>=2.6.0,<3",
	"deepdiff>=6.0.0,<7"
]

[tool.ruff]
line-length = 100
target-version = "py311"
